move cjson-related things to the separate package & add spot bug exclusion
