+++ src/main/java/com/github/fppt/jedismock/operations/scripting/Eval.java
@@ -3,6 +3,7 @@ package com.github.fppt.jedismock.operations.scripting;
 import com.github.fppt.jedismock.datastructures.Slice;
 import com.github.fppt.jedismock.operations.AbstractRedisOperation;
 import com.github.fppt.jedismock.operations.RedisCommand;
+import com.github.fppt.jedismock.operations.scripting.cjson.LuaCjsonLib;
 import com.github.fppt.jedismock.server.Response;
 import com.github.fppt.jedismock.storage.OperationExecutorState;
 import com.github.fppt.jedismock.storage.RedisBase;
